Added documentation for Bower
I added documentation for how to use this in a bower project, for both
css and sass. Bower doesn’t have one for less but it could be easily
added if someone needs it.

Added period and fixed word wrap

turned off word wrap in sublime and fixed the indentation

Fixed indentation

Added </p> and fixed indentation (again)