Change the Theme Primary Background to Primary Color
I noticed that changing the primary theme color by overriding the primary color variable doesn't affects some of the elements backgound color. This happens because the .bg-primary is using gray-900 variable rather than primary variable. Although the primary variable is pointing to the same grey-900 variable, but changing the value of the primary variable doesn't change the .bg-primary. This change (commit) fixed that, and makes changing the theme primary color (including bg-primary) easy by only overriding the value of a single variable (primary).